soc: octeontx2: Mbox communication support btw AF and it's VFs
Usually VFs attached to PFs other than AF do not communicate with AF
directly. Instead they are supposed to first send message to the PF
they are residing on and PF finally forwards it to the AF. Responses
to messages are handled in the reverse order.

On the other hand if VFs are on AF (PF0) itself then direct mailbox
communication is possible since there's no other PF in the way.

This patch addresses this particular case and adds support for
handling it.

Signed-off-by: Tomasz Duszynski <tduszynski@marvell.com>